+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
 
 - payment/drand: Add `library` feature
 - gateway: Use nois-payment dependency with `library` feature
+- Upgrade cosmwasm to 1.3.1
 
 ## [0.13.7] - 2023-07-01